requirements.txt
bidict==0.23.1
dnspython==2.6.1
eventlet==0.37.0
greenlet==3.1.0
h11==0.14.0
python-engineio==4.9.1
python-socketio==5.11.4
simple-websocket==1.0.0
wsproto==1.2.0
